requirements.txt
streamlit
requests
openai
bs4
google-generativeai
anthropic
mistralai
PyMuPDF
langchain_ollama